Risolvi tutti i warning del compilatore (-Wall -Wextra)
Warning sistemati (da 16 a 0):

1. -Wreorder in Game::Game():
   Riordinata la initializer list per corrispondere all'ordine di
   dichiarazione dei membri in Game.h (window, boss, state, gameMode,
   isRunning, currentLevel, selectedModeIndex, menuItemIndex,
   musicEnabled, lightningTimer, configJoyStep).

2. -Wsign-compare in Game.cpp (2 occorrenze):
   event.joystickButton.button e' unsigned int, config.joy_jump e' int.
   Aggiunto cast esplicito (unsigned)config.joy_jump nei confronti.

3. -Wsign-compare in Boss.cpp:
   shootTimer e' uint32_t, l'espressione (1500 - level*100) e' int.
   Aggiunto cast esplicito (uint32_t) al secondo operando.

4. -Wunused-but-set-variable in Game::update():
   Variabile locale 'moved' impostata ma mai letta. Rimossa del tutto
   (le setDirection() gia' comunicano l'input al player).

Build verificata: 0 warning, 0 errori.
Nessuna modifica al comportamento del gioco.

Branch: Ref_20260816

+++ src/Boss.cpp
@@ -63,8 +63,10 @@ void Boss::update(float playerX, float playerY, std::vector<Projectile>& bossPro
     if (pos.y < UI_HEIGHT + size/2 || pos.y > screenHeight - size/2) dy = -dy;
     shootTimer += 16;
 
-    // Sparo a ventaglio: il cooldown diminuisce col livello (min ~500 ms)
-    if (shootTimer > (1500 - level * 100)) {
+    // Sparo a ventaglio: il cooldown diminuisce col livello (min ~500 ms).
+    // Cast a uint32_t per evitare -Wsign-compare fra shootTimer (unsigned) e
+    // l'espressione (1500 - level*100) che e' int firmato.
+    if (shootTimer > (uint32_t)(1500 - level * 100)) {
         shootTimer = 0;
         float dxp = playerX - pos.x, dyp = playerY - pos.y;
         float dist = sqrt(dxp*dxp + dyp*dyp);
